break up wing file into separate files
