+++ 1682/gas_male/submodels/beam.py
@@ -1,0 +1,39 @@
+" discretized beam model "
+from gpkit import Model, Variable, vectorize
+
+class Beam(Model):
+    "discretized beam bending model"
+    def __init__(self, N, q, **kwargs):
+
+        with vectorize(N-1):
+            EIbar = Variable("\\bar{EI}", "-",
+                             "normalized YM and moment of inertia")
+
+        with vectorize(N):
+            qbar = Variable("\\bar{q}", q, "-", "normalized loading")
+            Sbar = Variable("\\bar{S}", "-", "normalized shear")
+            Mbar = Variable("\\bar{M}", "-", "normalized moment")
+            th = Variable("\\theta", "-", "deflection slope")
+            dbar = Variable("\\bar{\\delta}", "-", "normalized displacement")
+
+
+        Sbartip = Variable("\\bar{S}_{tip}", 1e-10, "-", "Tip loading")
+        Mbartip = Variable("\\bar{M}_{tip}", 1e-10, "-", "Tip moment")
+        throot = Variable("\\theta_{root}", 1e-10, "-", "Base angle")
+        dbarroot = Variable("\\bar{\\delta}_{root}", 1e-10, "-",
+                            "Base deflection")
+        dx = Variable("dx", "-", "normalized length of element")
+
+        constraints = [
+            Sbar[:-1] >= Sbar[1:] + 0.5*dx*(qbar[:-1] + qbar[1:]),
+            Sbar[-1] >= Sbartip,
+            Mbar[:-1] >= Mbar[1:] + 0.5*dx*(Sbar[:-1] + Sbar[1:]),
+            Mbar[-1] >= Mbartip,
+            th[0] >= throot,
+            th[1:] >= th[:-1] + 0.5*dx*(Mbar[1:] + Mbar[:-1])/EIbar,
+            dbar[0] >= dbarroot,
+            dbar[1:] >= dbar[:-1] + 0.5*dx*(th[1:] + th[:-1]),
+            1 == (N-1)*dx,
+            ]
+
+        Model.__init__(self, None, constraints, **kwargs)
