+++ 1682/gas_male/submodels/constant_taper_chord.py
@@ -1,0 +1,8 @@
+" constant taper chord "
+import numpy as np
+
+def c_bar(lam, N):
+    "returns wing chord lengths for constant taper wing"
+    eta = np.linspace(0, 1, N)
+    c = 2/(1+lam)*(1+(lam-1)*eta)
+    return c
